+++ src/main/java/server/transfer/consumer/Consumer.java
@@ -9,7 +9,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import server.transfer.data.ObservationData;
-import server.transfer.send.Sender;
+import server.transfer.sender.Sender;
 
 /**
  * Consumes data from Kafka
