+++ src/main/java/server/transfer/converter/GraphiteConverter.java
@@ -1,11 +1,11 @@
-package server.transfer.convert;
+package server.transfer.converter;
 
 import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.python.core.PyList;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import server.transfer.convert.util.PythonMetricUtil;
+import server.transfer.converter.util.PythonMetricUtil;
 import server.transfer.data.ObservationData;
 
 /**
